feat(fe): make component for search lang tags filter
